add working consumer

+++ src/Wsrc.Domain/Models/KickChatMessageSender.cs
@@ -4,7 +4,7 @@ namespace Wsrc.Domain;
 
 public class KickChatMessageSender
 {
-    [JsonPropertyName("id")] public long Id { get; set; }
+    [JsonPropertyName("id")] public int Id { get; set; }
 
     [JsonPropertyName("username")] public string Username { get; set; }
